Refactor and enhance API functionality across multiple services
- Updated campaign budget handling to ensure numeric type consistency in SQL queries.
- Improved Bedrock client configuration and added local development fallback for demo purposes.
- Enhanced wallet service tests to ensure accurate transaction handling and balance updates.
- Refactored wallet service to use UUIDs for shopkeeper references instead of Cognito IDs.
- Updated proxy routes for brand authentication to be more explicit.
- Improved API testing scripts for comprehensive coverage of authentication and brand functionalities.
- Added new test scripts for Bedrock models and brand APIs to ensure reliability and performance.
- Implemented error handling and logging improvements across various test scripts.
- Created new test scripts for Prisma database connection verification.

+++ src/lib/services/campaign-matcher.ts
@@ -120,9 +120,9 @@ export class CampaignMatcher {
       // Update budget
       const updateSql = `
         UPDATE campaigns
-        SET remaining_budget = $1,
+        SET remaining_budget = $1::numeric,
             status = CASE
-              WHEN $1 <= 0 THEN 'completed'
+              WHEN $1::numeric <= 0 THEN 'completed'
               ELSE status
             END
         WHERE id = $2
